@@ -1,8 +1,12 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <!--<LangVersion>preview</LangVersion>
+    <EnablePreviewFeatures>true</EnablePreviewFeatures>-->
+    <Version>1.0.1</Version>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
     <Authors>ShimakazeProject, frg2089</Authors>
     <PackageProjectUrl>https://github.com/ShimakazeProject/Shimakaze.Tools</PackageProjectUrl>
@@ -12,14 +16,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Shimakaze.Models.Csf" Version="1.0.0" />
+    <ProjectReference Include="..\..\..\Shimakaze.Models.Csf\Shimakaze.Models.Csf.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\..\Shimakaze.Tools.InternalUtils\ByteArrayExtensions.cs" />
-    <Compile Include="..\..\..\Shimakaze.Tools.InternalUtils\ByteUtils.cs" />
     <Compile Include="..\..\..\Shimakaze.Tools.InternalUtils\CsfUtils.cs" />
-    <Compile Include="..\..\..\Shimakaze.Tools.InternalUtils\SystemExtersions.cs" />
+    <Compile Include="..\..\..\Shimakaze.Tools.InternalUtils\EachExtension.cs" />
   </ItemGroup>
 
 </Project>
