@@ -1,22 +1,24 @@
-﻿namespace Shimakaze.Tools.Csf.Serialization.Csf
+﻿namespace Shimakaze.Tools.Csf.Serialization.Csf;
+
+internal static class Constants
 {
-    internal static class Constants
-    {
-        /// <summary>
-        /// ' FSC'
-        /// </summary>
-        public const int CSF_FLAG_RAW = 1129530912;
-        /// <summary>
-        /// ' LBL'
-        /// </summary>
-        public const int LBL_FLAG_RAW = 1279413280;
-        /// <summary>
-        /// ' RTS'
-        /// </summary>
-        public const int STR_FLAG_RAW = 1398034976;
-        /// <summary>
-        /// 'WRTS'
-        /// </summary>
-        public const int STRW_FLG_RAW = 1398035031;
-    }
+    /// <summary>
+    /// ' FSC'
+    /// </summary>
+    public const int CSF_FLAG_RAW = 1129530912;
+
+    /// <summary>
+    /// ' LBL'
+    /// </summary>
+    public const int LBL_FLAG_RAW = 1279413280;
+
+    /// <summary>
+    /// ' RTS'
+    /// </summary>
+    public const int STR_FLAG_RAW = 1398034976;
+
+    /// <summary>
+    /// 'WRTS'
+    /// </summary>
+    public const int STRW_FLG_RAW = 1398035031;
 }
