@@ -1,14 +1,15 @@
-namespace Shimakaze.Tools.Csf.Serialization.Json
+namespace Shimakaze.Tools.Csf.Serialization.Json;
+
+internal class Constants
 {
-    internal class Constants
+    public static class SchemaUrls
     {
-        public static class SchemaUrls
-        {
-            public const string BASE_URL = "https://shimakazeproject.github.io/Schemas";
-            public const string V1 = BASE_URL + "/json/csf/v1.json";
-            public const string V2 = BASE_URL + "/json/csf/v2.json";
-        }
-        public static readonly string[] LanguageList = new[] {
+        public const string BASE_URL = "https://shimakazeproject.github.io/Schemas";
+        public const string V1 = BASE_URL + "/json/csf/v1.json";
+        public const string V2 = BASE_URL + "/json/csf/v2.json";
+    }
+
+    public static readonly string[] LanguageList = new[] {
             "en_US",
             "en_UK",
             "de",
@@ -20,5 +21,4 @@ namespace Shimakaze.Tools.Csf.Serialization.Json
             "kr",
             "zh"
         };
-    }
 }
