@@ -1,13 +1,14 @@
-namespace Shimakaze.Tools.Csf.Serialization.Xml
+namespace Shimakaze.Tools.Csf.Serialization.Xml;
+
+internal class Constants
 {
-    internal class Constants
+    public static class SchemaUrls
     {
-        public static class SchemaUrls
-        {
-            public const string BASE_URL = "https://shimakazeproject.github.io/Schemas";
-            public const string V1 = BASE_URL + "/xml/csf/v1.xsd";
-        }
-        public static readonly string[] LanguageList = new[] {
+        public const string BASE_URL = "https://shimakazeproject.github.io/Schemas";
+        public const string V1 = BASE_URL + "/xml/csf/v1.xsd";
+    }
+
+    public static readonly string[] LanguageList = new[] {
             "en_US",
             "en_UK",
             "de",
@@ -19,5 +20,4 @@ namespace Shimakaze.Tools.Csf.Serialization.Xml
             "kr",
             "zh"
         };
-    }
 }
