@@ -1,14 +1,10 @@
-
-using System;
-using System.Linq;
 using System.Xml;
-using System.Xml.Serialization;
 
-namespace Shimakaze.Tools.Csf.Serialization.Xml.V1
+namespace Shimakaze.Tools.Csf.Serialization.Xml.V1;
+
+public interface IXmlConverter<T>
 {
-    public interface IXmlConverter<T>
-    {
-        T Deserialize(XmlReader reader);
-        void Serialize(XmlWriter writer, T value);
-    }
+    T Deserialize(XmlReader reader);
+
+    void Serialize(XmlWriter writer, T value);
 }
